refactor: make oauth callback reuseable (util function)
