+++ src/config/passport.ts
@@ -1,7 +1,7 @@
 import { User } from "@/data/entities/User";
 import repository from "@/data/repositories";
-import { IOAuth2Command, OAuth2UseCase } from "@/use-cases/OAuth2UseCase";
 import { SigninUseCase } from "@/use-cases/SigninUseCase";
+import { oauth2Callback } from "@/utils/passportUtils";
 import { plainToInstance } from "class-transformer";
 import passport from "passport";
 import { Strategy as FacebookStrategy } from "passport-facebook";
@@ -47,29 +47,7 @@ passport.use(
       scope: ["profile", "email"],
       passReqToCallback: true,
     },
-    async (req, accessToken, refreshToken, profile, done) => {
-      const oauth2UseCase = new OAuth2UseCase(repository);
-      try {
-        const id = profile.id;
-        const displayName = profile.displayName;
-        const email = profile.emails?.[0]?.value || null;
-        const imageUrl = profile.photos?.[0]?.value || null;
-
-        const command: IOAuth2Command = {
-          providerName: "google",
-          id,
-          displayName,
-          email,
-          imageUrl,
-        };
-
-        const user = await oauth2UseCase.execute(command);
-
-        return done(null, user);
-      } catch (error) {
-        return done(error);
-      }
-    },
+    oauth2Callback,
   ),
 );
 
@@ -84,29 +62,7 @@ passport.use(
       scope: ["email", "public_profile"],
       passReqToCallback: true,
     },
-    async (req, accessToken, refreshToken, profile, done) => {
-      const oauth2UseCase = new OAuth2UseCase(repository);
-      try {
-        const id = profile.id;
-        const displayName = profile.displayName;
-        const email = profile.emails?.[0]?.value || null;
-        const imageUrl = profile.photos?.[0]?.value || null;
-
-        const command: IOAuth2Command = {
-          providerName: "facebook",
-          id,
-          displayName,
-          email,
-          imageUrl,
-        };
-
-        const user = await oauth2UseCase.execute(command);
-
-        return done(null, user);
-      } catch (error) {
-        return done(error);
-      }
-    },
+    oauth2Callback,
   ),
 );
